==> Lesson_5_Gas_Analysis/0. assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd
import numpy as np

df_08 = pd.read_csv('all_alpha_08.csv')
df_18 = pd.read_csv('all_alpha_18.csv')

### Find the correct count for the 2008 dataset.

# number of samples
df_08.shape

# duplicated rows
df_08[df_08.duplicated(keep='first')].count()

# rows with missing data
df_08[df_08.isnull().any(axis=1)].count()

### Find the correct count for the 2018 dataset.

# number of samples
df_18.shape

# duplicated rows
df_18[df_18.duplicated(keep = 'first')].count()

# rows with missing data
df_18[df_18.isnull().any(axis=1)].count()

df_08.dtypes

df_18.dtypes

# chec unique values
df_08.nunique()

df_08.groupby(['Fuel']).mean().Displ

df_18.groupby(['Fuel']).mean().Displ

## QUESTION 3 OF 4
Which of the following features need to be converted to floats from strings?
