==> Lesson_5_Gas_Analysis/3. inspect_datatypes.ipynb <==
# Inspecting Data Types
Use the space below to explore `data_08.csv` and `data_18.csv` to answer the quiz questions below regarding datatypes.

import pandas as pd
df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

df_08.head()

df_18.head()

## QUESTION 1 OF 4
Which of the following changes could be made to make the `cyl` column in both datasets clear and consistent?

df_08['cyl'].head()

# replace cyl in 08 dataset as int.
x = df_08.cyl.str.extract('(\d+)')
x = pd.to_numeric(x[0][:])
df_08['cyl'] = x
df_08.head()

# replace cyl in 18 dataset as int.
df_18['cyl'] = pd.to_numeric(df_18['cyl']).astype(int)

# save the new datatype into csv
df_08.to_csv('data_08.csv', index=False)
df_18.to_csv('data_18.csv', index=False)

## QUESTION 2 OF 4
Which of the following changes could be made to make the `air_pollution_score` column in both datasets consistent?

import pandas as pd
df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

df_08.dtypes

df_18.dtypes

# convert strings to floats in the 2008 column.
# df_08['air_pollution_score'] = pd.to_numeric(df_08['air_pollution_score']).astype(floats)
# df_18['air_pollution_score'] = pd.to_numeric(df_18['air_pollution_score']).astype(floats)
# pd.to_numeric(df_08['air_pollution_score'][0][:])
# df_08['air_pollution_score'].nunique()
df_08.groupby(['air_pollution_score']).mean().cyl
