==> Lesson_5_Gas_Analysis/10. results_merged.ipynb <==
# Results with Merged Dataset
Remember to use your new dataset, `combined_dataset.csv`

# load dataset
import pandas as pd

df = pd.read_csv('combined_dataset.csv')

### Q5: For all of the models that were produced in 2008 that are still being produced now, how much has the mpg improved and which vehicle improved the most?

df.dtypes

model_mpg_08 = df.groupby('model_2008').cmb_mpg_2008.mean()

model_mpg_18 = df.groupby('model_2018').cmb_mpg_2018.mean()

(df.model_2008 == df.model_2018).all()

mpg_change = model_mpg_18 - model_mpg_08

mpg_change[mpg_change==mpg_change.max()].index[0]
